Fixed a reference issue for hazards and creatures
The temporary components were not copied, and as such, when a creature/hazard was deleted from the temporary object, it was also deleted in the original.
Other similar issues probably existed.
Fixed by making a hard copy of the original (using the same serialization function that are used for saving to the local storage)